manejo de errores

+++ gastroticket-backend/src/main/java/uoc/tfm/gastroticket/empresas/service/EmpresasService.java
@@ -2,7 +2,6 @@ package uoc.tfm.gastroticket.empresas.service;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,6 +12,7 @@ import io.jsonwebtoken.io.IOException;
 import jakarta.transaction.Transactional;
 import uoc.tfm.gastroticket.empresas.model.EmpresasDTO;
 import uoc.tfm.gastroticket.empresas.repository.EmpresasRepository;
+import uoc.tfm.gastroticket.user.Role;
 import uoc.tfm.gastroticket.user.User;
 import uoc.tfm.gastroticket.user.UserRepository;
 
@@ -34,18 +34,33 @@ public class EmpresasService {
     }
 
     public void createEmpresa(String nombre, String email) {
+        User _user = new User();
+        _user.setUsername(email);
+        _user.setRole(Role.EMPRESA);
+        _user = userRepository.save(_user);
+
         EmpresasDTO _empresa = new EmpresasDTO();
         _empresa.setNombre(nombre);
         _empresa.setEmail(email);
+        _empresa.setUserId(_user.getId());
         empresaRepo.save(_empresa);
     }
 
     @SuppressWarnings({ "unchecked", "rawtypes" })
     public ResponseEntity<?> editarEmpresa(long id, String nombre, String email) {
-        EmpresasDTO _empresa = empresaRepo.findById(id).get();
-        _empresa.setNombre(nombre);
-        _empresa.setEmail(email);
         try {
+            EmpresasDTO _empresa = empresaRepo.findById(id).get();
+            if (_empresa.getUserId() != null) {
+                User user = userRepository.findById(_empresa.getUserId()).orElse(null);
+                if (user == null) {
+                    throw new RuntimeException("Usuario no encontrado");
+                }
+                user.setUsername(email);
+                userRepository.save(user);
+            }
+
+            _empresa.setNombre(nombre);
+            _empresa.setEmail(email);
             empresaRepo.save(_empresa);
         } catch (IOException e) {
             return new ResponseEntity("Error al editar la empresa", HttpStatus.BAD_REQUEST);
@@ -54,11 +69,23 @@ public class EmpresasService {
                 HttpStatus.OK);
     }
 
-    public void eliminarEmpresa(long id) {
-        EmpresasDTO empresa = empresaRepo.getReferenceById(id);
-        User user = userRepository.findById(empresa.getUserId()).orElse(null);
-        if (user != null)
-            userRepository.delete(user);
-        empresaRepo.deleteById(id);
+    @SuppressWarnings({ "rawtypes", "unchecked" })
+    public ResponseEntity<?> eliminarEmpresa(long id) {
+        try {
+            if (empresaRepo.existsById(id)) {
+                EmpresasDTO _empresa = empresaRepo.findById(id).get();
+                if (_empresa.getUserId() != null) {
+                    User user = userRepository.findById(_empresa.getUserId()).orElse(null);
+                    if (user != null) {
+                        userRepository.delete(user);
+                    }
+                }
+                empresaRepo.deleteById(id);
+            }
+        } catch (IOException e) {
+            return new ResponseEntity("Error al eliminar la empresa", HttpStatus.BAD_REQUEST);
+        }
+        return new ResponseEntity(Collections.singletonMap("mensaje", "La empresa se ha eliminado correctamente"),
+                HttpStatus.OK);
     }
 }
